perf(extension): load the compiled entrypoint directly
